requirements.txt
# GitHub API
PyGithub>=1.55

# OpenAI client
openai>=0.27.0

jinja2 >=3.1.2

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-env >=0.6.2
flake8 >=6.0.0
black >=23.3.0